requirements.txt
lightning
soundstream
wandb
torchmetrics
pesq
scikit-learn
soundfile
torchmetrics[audio]
micromind
torchsummary